iOS reader/holder now support L2CAP.
This is a fairly large change; both the holder and the reader have been
extended to support L2CAP.  In service of this, the original flow has been
refactored into more explicit state machines, and the L2CAP flow has been
added to those state machines.

Both the reader and holder have `useL2CAP` booleans; if either is `false`,
the reader and holder will use the old flow to communicate.  This should
also mean that they will use the old flow when working with other readers
or holders that do not support L2CAP.

Some of this work (notably the `*Connection.swift` files) is derived from
Paul Wilkinson's MIT-licensed L2Cap library:

https://github.com/paulw11/L2Cap

This repo is MIT-licensed as well, so the licenses are compatable.

We aren't using the L2Cap library as-is for a variety of reasons, but the
main reason is that the behaviour we need differs significantly from the
behaviour L2Cap (the library) offers.

There are a couple of places in this change that are hacking around
impedence mismatches:

- we seem to need to have `notify` on the L2CAP characteristic in order
  to propagate the PSM to central, though 18013-5 Annex A claims the
  only required property is `read` -- we're not out of spec, since we're
  offering an optional property, but it remains to be seen how that will
  interact with 3rd party centrals

- 18013-5 specifies no framing for the request or response sent over the
  L2CAP stream, so we have to infer when the data has arrived from timing;
  this seems like a fragile method particularly if confronted by noisy
  radio environments